api_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buildstate-api"
version = "1.0.0"
description = "Build State API Service - FastAPI service for managing multi-cloud IaaS image build states"
authors = [
    {name = "Build State Team", email = "buildstate@example.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "PyJWT==2.8.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
    "httpx==0.25.2",
    "sqlalchemy>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[tool.setuptools]
packages = ["app", "app.core", "app.routers"]

[tool.setuptools.package-dir]
"" = "."